// src/common/enums/enums.ts
export enum enumOsType {
  ANDROID = 'android',
  IOS = 'ios',
  WINDOWS = 'windows',
  WEB = 'web',
}

export enum enumVersionPriority {
  OPTIONAL = 'optional',
  NECESSARY = 'necessary',
}

export enum enumPriority {
  HIGHT = 'hight',
  MID = 'mid',
  LOW = 'low',
}
// ...
export enum enumTypeOfActivity {
  USER = 'user',
  VERSION = 'versions',
  GALLERY = 'gallery',
  ROLE = 'role',
  RATE = 'hc',
  PERMISSION = 'permission',
}